Fix #3: Make the application runnable inside the container
Added dependency on 'spring-boot-starter-web'.
Implemented the new 'ServletInitializer' class.
Changed absolute paths to relative.
Updated project package name.
Changed the application class name to 'PowerAuthWebAuthApplication'.
Removed unused property from 'application.properties' file.
Removed static imports.
Changed website title.
