+++ powerauth-webauth/src/main/java/io/getlime/security/powerauth/app/webauth/ServletInitializer.java
@@ -4,8 +4,8 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
  * http://www.apache.org/licenses/LICENSE-2.0
+ *
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,33 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.lime.webauth.model.entity;
+package io.getlime.security.powerauth.app.webauth;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.support.SpringBootServletInitializer;
 
 /**
- * @author Roman Strobl
+ * @author Petr Dvorak, petr@lime-company.eu
  */
-public class RegistrationMessage {
+public class ServletInitializer extends SpringBootServletInitializer {
 
-    private String action;
-    private boolean performUITest;
-
-    public RegistrationMessage() {
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(PowerAuthWebAuthApplication.class);
     }
 
-    public RegistrationMessage(String action, boolean performUITest) {
-        this.action = action;
-        this.performUITest = performUITest;
-    }
-
-    public String getAction() {
-        return action;
-    }
-
-    public boolean getPerformUITest() {
-        return performUITest;
-    }
-
-    public String toString() {
-        return "action: " + action + ", performUITest: " + performUITest;
-    }
 }
