+++ powerauth-webauth/src/main/java/io/getlime/security/powerauth/app/webauth/configuration/WebSocketConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.lime.webauth.configuration;
+package io.getlime.security.powerauth.app.webauth.configuration;
 
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.stereotype.Component;
