--- powerauth-webauth/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.data.rest.base-path=/api
