+++ powerauth-webauth/src/main/resources/templates/index.html
@@ -2,8 +2,8 @@
 <html>
 <head lang="en">
     <meta charset="UTF-8"/>
-    <title>Web Auth from Lime</title>
-    <link rel="stylesheet" href="/main.css"/>
+    <title>PowerAuth 2.0 Web Auth</title>
+    <link rel="stylesheet" href="./main.css"/>
 </head>
 <body>
 
@@ -11,7 +11,7 @@
     <div id="react"></div>
 </div>
 
-<script src="built/bundle.js"></script>
+<script src="./built/bundle.js"></script>
 
 </body>
 </html>
